requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.1
django-baton==5.0.0
django-modeltranslation==0.19.16
django-rosetta==0.10.2
django-widget-tweaks==1.5.0
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
polib==1.2.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
